[IMP] web: trigger an event when webclient is ready
With the actual structure of the webclient (action_manager, view_manager),
there is no easy way to decide whether or not a view is attached in the DOM,
because every rendering is performed in a detached element. For instance,
when a view_manager instantiates a new view, it appends it in its $el once
the view is ready, but it can't know if its $el is attached in the DOM (it is
only the case if the view_manager is the current action of the action manager,
and if the action manager's $el is in the DOM as well).

It happens we need to know when the webclient is totally loaded, i.e. when the
view has been loaded and attached in the DOM (e.g. scroll position restoring,
tips...).

This rev. introduces utility functions to centralize the appending/prepending of
content, so that an event can be triggered (at a unique place) when the content
is really in the DOM. These functions are used by the webclient, the action_manager
and the view_manager.

Another change is that the webclient (resp. the action_manager) keeps the action_manager
(resp. the view_manager) informed when they are attached in/detached from the DOM,
through the variable is_in_DOM.