[ADD] stock_barcode: editable list of products to set their barcodes
Editable list of product.product where the only field not to be
readonly is barcode. In an editable list, a carriage return sets
the focus in the first focuseable field of the following line. Since
most barcode scanners input the barcode followed by a carriage return,
this allows to fill in all product's barcode using just a scanner.