[IMP] web: restore scroll position
Simplification of the solution to restore the action manager's scroll position
by using the new event 'DOM_updated' triggered each time new content is attached
in the DOM.

The logic is now concentrated at one place, instead of duplicated in the
action_manager and view_manager.

Moreover, the action_manager is now able to restore its scroll position after
having been toggled by the webclient as well, which was not the case with the
former solution.