[IMP] web: scroll to the top on page changed
use case: list and kanban views

When clicking on the pager to skip to next/previous page of a list view, the
expected behavior is that the scroll position go back to the top. When editing
the pager (e.g. changing the limit), however, the scroll position should
not change.

For example, records 1-20 are displayed, i scroll to the bottom and edit the
pager to display record 1-40, i want the scroll position to stay unchanged.