[FIX] web: on_hashchange: update menu if inner_action exists
Sometimes, after calling do_load_state() on the action_manager, its inner_action
is undefined. For instance, this is the case when trying to open an url in which
an unknown action is specified. This is also the case when clicking on 'add a translation'
in the website as it goes to the backend (in which no inner action is loaded), to
open a modal.

This rev. prevents the webclient to crash when trying to access the action_descr
of an undefined action.