[IMP] web: clear url in app switcher
When refreshing in the app switcher, it should remain in the app switcher. This
rev. clears the url on the app switcher and restores it on back button click.

The private variable _ignore_hashchange has been introduced so that we don't
listen on the hashchange event triggered when clearing the url (as it would
again call toggle_app_switcher()).