[IMP] web(_tests): FieldMany2One: autocomplete
Since every rendering is now performed in fragments detached from the DOM, the
selectors used for the appendTo option of jquery-ui autocomplete don't match
anything. Worse, if they did match, FieldMany2One would be broken because they
would match DOM elements that will be detached as soon as the current view is
ready.

When the appendTo option is not specified, the autocomplete widget is appended
to the body, which is also the case when the specified selector doesn't match
anything.

This rev. thus removes the option appendTo, so that the autocomplete widgets
are always appended to the body. Also updates a tour's selector matching an
autocomplete ul as it isn't inside the .modal anymore.