[FIX] web: prevent crash when connection is lost/resumed
When connection is lost, a notification is displayed.  When the
connection is resumed, that notification is supposed to be destroyed.

However, the display method in notification manager returned a deferred,
so the web client was trying to destroy a deferred, which does not
work.  This commit makes sure that the web client destroys the widget,
not the deferred.