[FIX] web: view_manager: update control panel on title change
only if the control panel if visible.

This fixes a bug on the settings view (form view), on which the control panel
is hidden (the bar containing 'Apply' and 'Cancel' is unfortunately not a true
control panel). When clicking on 'Apply', for some weird reason, the event
'change:title' is triggered and the control panel is updated with the new title
('New'). However, as the flags hidden isn't mentionned, the control panel
visibility is toggled.