[TEMP] store estimated delivery price on sale order
Functional: When we edit a sale order and we use an external shipping
provider, the estimated delivery price is refreshed at each loading of
the SO. This is quite slow as the code will try to contact the external
provider at each refresh.

Technical: Delivery_price is already a non-stored computed field of
sale.order, BUT we are working in stable saas-6, so we cannot "simply"
add a store=True in the existing code.

!! TO BE MOVED TO MASTER odoo/odoo !!