[FIX] web: List view record reloading deferred
When a list view record is reloaded, there is a rendering lag because of
the addition/suppression of a table row. Since the list view editable need
to wait for the end of this rendering lag before computing the position of
the editable fields, a 100ms delay is added at the end of the reload_record
function.