[IMP] web: form_view sheet layout without chatter
The sheet grey background should take height 100% when there is no chatter in
the bottom, but it should only take the size of the sheet when there is one,
and the chatter should take 100% of the space left.

Moreover, in mobile, we override the .container rules applied on the chatter
as it should take full width.