[FIX/IMP] web: data_export: save exports
This commit fixes a bug when trying to save an export list. It also improves
a lot the style of this part of this modal, which has probably been left
behind in the new webclient refactoring.

Basically, DOM simplification, no more inline style, no more id selectors, but
only for the save exports part of the modal.

The code of data_export is ugly, and should definitely be refactored.