[FIX] web: date widget, set date when shown
In 36146b3 setting the date of the datetimepicker was done when it was
opened, but only for widget datetime. Thus we could get an issue for a
widget date when we wanted to use it to select two same date.

e.g if we select 23 september for a date, it works. if we then use the
widget on another date, 23 september is selected and it will not trigger
a value change if we leave this value.

This fix adds the date setting when opening the date widget, and also
prevents the bootstrap datetimepicker feature* of closing the widget
when setting the value of the date-only picker.

closes #8506
opw-648619

* this feature is explained in the following issue:
  Eonasdan/bootstrap-datetimepicker#603