[FIX] web: restrict width of sequence column
By default, the table layout algorithm tries to distribute column widths
relatively evenly. When there is no content whatsoever in the
table (outside of the column titles) the sequence column gets a lower
"weight" than others but still some, and is much wider than the
~1-character it gets once other columns actually have content (and it
still has none).

Force-constraint it to a width of 1 (smallest possible width, forcefully
enlarged to hold a resize character thing which is used in the actual
rows), limits reflow incurred when actual content gets added to tables.