[FIX] project_timesheet_synchro : Fix a bug with select2 that occurs on android native browser.
* The timesheet app on android uses the native web browser.
With this browser, there is a bug with select2 dropdowns which closes immediatly after opening,
because of the mobile virtual keyboard popping up. This commit fixes this bug by editing the select2.js file.
Since this file is part of a library, we copy it from web into this module, so that the fix is
only applied to the mobile and chrome extension versions of the app and not the whole web client.