[FIX] web: prevent logging of bus longpolling errors
The longpolling have been disabled in test mode, to prevent timeouts
(test mode only has one cursor, and long polls keep it for 60s, which is
not good).  This commits prevent the client from logging the resulting
error in the browser log, because that is how phantomjs determines if
there is an error.