[FIX] web: formview oe_title layout in mobile
The title went below the avatar in mobile. The reason was the edit mode
could be sometimes ugly but here the readonly will always look better
for less rules.

This commit fixes also the title sizes in mobile which were not correct
since the LESS max function does not work (the rules are better now
anyway).

+ Break words in oe_title in mobile