[ADD] web: new panel for expiration database
With our new ping process deployed on internal, each database will receive an expiration date and an expiration reason when pinging our server.

The app switcher will display a notice 30 days before that expiration date to allow the customer to correct the situation according to the expiration reason.

The current possible reasons are:
- upsell: the customer has more users than specified in the contract
- renewal: the contract reaches its normal ending date and should be renewed
- trial: the trial period of 30 days is reaching its end

The possible actions redirect to our server with the route depending on the action needed.

Courtesy of @adr-odoo, @mge-odoo and @dbo-odoo