[FIX] web: action_manager: restore internal state on rejected action
When an error occured while executing an action (e.g. access rights), the
action manager's state was inconsistent as inner_action, inner_widget and
action_stack were all updated before trying to execute the action. Then, if the
user tried to open another action, the previous action wasn't properly detached
as the action manager lost its reference. As a result, the new action was
displayed over the old one.

We now restore inner_action, inner_widget and action_stack to their previous
state in the case of a rejected deferred.

A better fix would be to update all three variables once the deferred is
resolved, but it would require a refactoring of the way the action manager
and the view manager communicate to perform the do_push_state and to compute
the breadcrumbs. Indeed, those operations directly rely on the internal state
of the action manager, and they are performed before the callbacks in
push_state are called. Such a refactoring should probably not be done in
stable.

Partly fixes issue #8886 (odoo/odoo).