[FIX] l10n_us_check_printing: amount paid in invoice
The amount paid is not correctly calculated since invoice_id is not
always filled in an account move line. The reason is that a payment can
be used to pay several invoices, therefore the field is often kept empty
(see for example method _create_payment_entry of account.payment.

In order to find the amount paid with a check for a given invoice, we
use the payment reconciliation table.

opw-652413