[FIX] delivery_fedex: 7000 Unable to obtain courtesy rates
When sending a delivery order to Fedex,
it sometimes happens that the Fedex server
doesn't reply the shipping cost, and consider this
as a warning `7000 Unable to obtain courtesy rates`.

Mostlikely, it only happens with the Fedex test server.
When retrying to send a request that failed with this warning,
without changing anything, it worked.

A blocking error should return a
`HighestSeverity` error anyway.

opw-652534