[IMP] website_sign: iframe resizing method
Before this commit the pdf iframe was resized according to the window
thanks to a resize event handler. This caused some problem on mobile as
opening the keyboard triggered this resize event handler too.

Solution is to use flex layout to automatically have a nicely resized
iframe (this solution is already in use on odoo v9).