[FIX] web: globe translation icon and notebook
The position of the globe icon corresponding to its transtable field was
set when the view form start being edited.

When such an icon was contained by a not displayed notebook page tab,
wrong pixel measurement could be used thus leading to an error in the
computed  icon position.

This commit replace this with static positioning and a negative
margin-left.

closes #104
opw-652972