[FIX] website_sign: sign items DOM detach by PDF viewer
The PDFJS viewer is modifying the DOM while going trough a long PDF.
It replaces the invisible pages content by a loading icon after
scrolling, zooming and many other actions in the PDFJS interface. The
solution that was implemented was not sufficient enough and introduced
other problems in mobile (such as loosing focuses or keyboard).

The solution here is to check at regular interval if the signature
items are still attached correctly and only reappend the lost ones.
This also permits to reappend lost ones on key moments (before checking
completion, going to the next field, ...). This allows to always have a
correct state, to never lose focus on an already attached item and to
always have all the fields in the DOM on the moment the system checks
if everything is completed.