[FIX] delivery_fedex: Rate Service call Failed attempting to obtain shipment level rates
This revision is related to 897bbc0be458ad59d1fa32e736cde9a20dfc2acd

Sometimes, Fedex cannot obtain the currency rates. Maybe only when
this is a test server. And in such case, Fedex doesn't return
a price for the delivery request. The below warning are returned
by Fedex:

7000: Unable to obtain courtesy rates.
7005: Rate Service call Failed attempting to obtain shipment level rates

As these errors are labelled as `Warnings`, and not as blocking issues,
the process keeps going. At some point, the code relies on the fact
the price is returned by Fedex, either in the currency asked,
and if not, in USD, while it's not in this case.

For instance, in `fedex_send_shipping`.

So, if Fedex is unable to obtain the currency rates, and doesn't
return a price, we force the price, at 0.00 `USD`.

opw-653040