[FIX] web: do not display loading error message
In several situations, loading a view can be forbidden server side. The
most common situation is the user access restriction. It can also be the
case with the use of method `view_init`, which can raise an error (see
model crm.lead2opportunity.partner). In these situations, the server
properly handles the request of the user and displays the appropriate
error message. There is therefore no need of an additional "view
couldn't be loaded" message, which can be misleading.

opw-653331