[FIX] account_sepa: depends on invoicing, not accounting.
This revision is related to f93e99ff0b4b657a5045da71a1b72940a3e038c0.

`account_sepa` is installed by default for the localisation
of Europeans countries.

It must be possible to only install invoicing for these countries.
The full accounting `account_accountant` must not be a mandatory
dependence for these localisations.

The features of this module won't be reachable without
`account_accountant`, without the menus this module adds,
but this isn't important, as SEPA is meant for purchases,
while the Invoicing module is meant for sales only.

opw-653271