[FIX] website_version: don't disable save when no version to save
The "Save" button on a page can be used to save a modification of an
ir.ui.view, but also change in fields values.

There is currently no versionning of field values, so the versionning
save buttons where disabled when a field value is changed, but at the
same time the normal "Save" button was sometimes also disabled whilst it
should not have been.

With this commit:
- only the versionning save buttons are disabled when no .o_dirty
  element is found,
- the "Save" button keep its normal behaviour.

opw-654277