[FIX] web: dumb workaround for event handlers not being triggered
The app switcher event handlers related to the expiration panel
arent't triggered when the blockUI is active, probably due to
the event delegation to the parent widget element,
which is blocked by blockUI.
This makes users unable to actually submit their registration
code or buy a subscription.

Quick temporary fix until a cleaner one is found.