[FIX] web: list_view: empty x2manylist and onwrite list editable
Commit 6ef43c7 correctly displays nocontent helper on empty list views, instead
of rendering the table itself. Functionnally, it solves a problem as the helper
is now correctly displayed when needed. Technically, it also looks like an
improvement as the empty table isn't rendered at all. Unfortunately, this
commit breaks 2 tests.

First, the widget_x2many tour doesn't pass anymore because x2many list views
shouldn't display this helper, but rather the 'Add an item' button inline, so
in this case, the table should be rendered.

Second, the onwrite mechanism of the list editable requires the table to be in
the DOM even if it is empty, so that it is able to insert a first record.

This commit fixes these two issues.