[FIX] web_gantt: improve horrible hack with the scroll position
There's a dirty hack in the gantt view because it cannot be rendered
outside of the DOM: we render it somewhere in the DOM then, once the
webclient is loaded, we move it inside the action manager. After the
upgrade to dhtmlxgantt 3.3 pro, the scroll positions were lost after
the move.