[FIX] web: display data in list view when grouped by default
The issue is: when a list view is opened in grouped mode by default, it
displays no content, but if we remove and reapply the groupby, it
displays the content.

The list view worked before by chance, and was broken recently.  The
reason is that we check if the dataset is empty to decide if we want to
display the no content helper.  However, the dataset is actually empty
when we start the view grouped, even if we have the data.

The only real fix I can see is to rewrite this pile of c**p at some
point, but meanwhile, this fix will do.