[FIX] delivery_ups: remove demo data
This is not really a good idea to make public the developer's UPS
account details. In this case for example, the Belgian account has been
blocked by UPS, leading to an error which is not very clear ("The
Shippers shipper number cannot be used for the shipment").

The Belgian UPS default service type is also set to 'UPS Standard' since
'UPS Ground' is not available in Belgium.

opw-657536