[FIX] web: properly remove nvd3 tooltips
with the recent update of nvd3 to the last version, we fixed a lot of
issues with their tooltips, but we also introduced at least a new
problem: the tooltips were not removed when the graph is detached from
the dom.

This commit sadly simply remove all tooltips when the view is detached,
to prevent them from clogging up the UI.  I'd like to see a better
solution.