[FIX] web: list view: allow to sort columns even if grouped
The sort feature had been disabled at rev. 7ab5075 in grouped mode because it
is disastrous, honestly (the groups are sorted in an unpredictable order,
unfolded groups are automatically folded...), and we can't really do something
to improve it given the current implementation of the list view.

Despite everything, this *feature* is popular so, as expected, we have to
re-introduce it.