[FIX] web: translate icon rendering
In fact, since commit 4214826ee75e0889f7bcb7fe41d88b4809a9b74a the position relative on icon parent is
not useful anymore since the icon is not absolute positioned anymore.

Moreover previous fix was breaking the listview editable.

However, the icon width had to be set to not break alignements of
translatable fields in a .o_row environment (and to place the icon
at the correct position).