[FIX] l10n_us_check_printing: fix element positions
* don't cache one pdf for three printing types (top, middle, bottom),

* fix positions for top and middle printing with a fixed 8.5in x 11in
  letter paper size,

* use zoom to avoid getting margin.

The bigger issue this commit fixes, is that margins of unknown sizes
were added by wkhtmltopdf. It seems that even if we have everything on
8.5in x 11in and create a pdf with 8.5in x 11in size, the snapshot taken
from qtwebkit doesn't have a 8.5in x 11in viewport.

The solution used, is sizing everything for a fixed 8.5in x 11in (us
letter size) and then we zoom so the content gets up to the qtwebkit
virtual full width size. Then when printing to 8.5in x 11in the bigger
content is scaled back to the original size.

There seems to be other solutions (zoom, viewport width and other
wkhtmltopdf options) but nothing seems to work for example on a
non-patched-for-wkhtmltopdf qt on linux.

The zoom used has been choosen empirically, it could depends on multiple
factors (operating system, wkhtmltopdf, qtwebkit, qt, resolution) and
need to be set bigger for an unexpected configuration.

closes #170
opw-652385