[IMP] account_contract_dashboard: clarify welcome message
The welcome message on the Revenue KPI dashboard saying it will be
removed after the first invoiced contract may not accurate enough.

The invoiced subscription must have a set "Deferred Revenue Category"
which will be set as account.asset.category on the invoice lines.

To frighten less people, this commit simply specify in the message that
the "contract" must be "recurrent".

first part of:
opw-658823