[FIX] web: translate icon position in listview editable
Since commit 4214826ee75e0889f7bcb7fe41d88b4809a9b74a the translate icon position is not absolute-computed
anymore. This made the translate icon disappear in listview editable
as every field is absolute-positioned (so the translate icon cannot be
correctly static-positioned with a negative margin-left).

The solution is to absolute-position the translate icon of a field which
is in a listview editable at the same time this field is absolute-positioned.