[ADD] stock_barcode: link to demo barcodes sheet in the 'dashboard'
When stock_barcode is installed with demo data, display on the 'dashboard' a
message with a link to the demo barcodes sheet. Upon closing this message (it
actually is a bootstrap alert) the user is given a "don't show this message
again" option.

Incidentally, used rem instead of em to define margins with a consistent unit.
(IE supports it since V8)