[FIX] inter_company_rules : misplaced sudo to prepare value to create PO/SO lines
The initial problem is when generating a PO from SO (or the inverse) the taxes were not propagated to the new created line. Taxes are company dependent field : so when we create a PO in company A, the PO line contains the taxes of company A, but we want the corresponding SO line in company B to have the taxes of company B, which wasn't the case.
This commit makes the taxes propagation works. The initial line is now browse as sudo() to get all the taxes related to this line (or its product). Then we filter them to keep only the ones from the right company.

This is a cherry-pick of commit odoo/odoo@b6fa479000f7afdcaf4b4a48dbf9db9fc6bf9002