[FIX] sale_contract: fix mismatch in signature for get_fiscal_position call
In saas-6, get_fiscal_position took the company_id
as its first argument; this is no longer the case
in v9 but this code wasn't updated accordingly.

This introduced bugs when trying to invoice
some subscriptions: the code tried to browse the
partner which had the id of the company; in some
cases the partner existed (but was not the correct
one), in other cases it didn't and crashed if the
correct partner had no fiscal position set (non-existing
record).