[FIX] account_reports: create missing translations
When creating a financial report, a menu and an action are automatically created.
This request duplicates the translation of the report for the action and menu.

The creation of the new translations can not be done during the creation of the
menu and action because:
- the translations are not loaded yet for reports created at the installation of
  a module (.po files loaded at the end)
- it would not fix the missing translations on the existing customers, with this
  fix, reloading the translations is enough to create the missing

Using a JOIN on ir.translation ensure the duplication is only done for missing
terms and should not deterioriate performances on big databases.

Requeries two query as needs to create translations for actions and menus.
The performance impact should be negligable in comparison of loading .po files.