[REF] website_sign: remove website dependency and custom dashboard
This commit regroups 3 big changes in features for website_sign:

* The website dependency has been removed. Signers now land on a custom
"frontend" page which does not require the odoo users to build a
website.

* The custom dashboard view which regrouped the template and request
models has been splitted into two standard kanban views. This allow
searches, grouping, filtering in a more standard and easy way.

* Backend users can now sign their documents without having to check
their mailbox to go to the frontend page. If the creator of a
signature request is part of the document signers, he is thus only one
click away of the signing page after creating it.