[FIX] web: reimplement autosize for extra fun (and some bugfixing)
autosize does not work correctly (it moves the screen in all directions
sometimes when the user press enter in the bottom of the screen in some
browsers)

So, we decided to have our own version, because of course, we can do it
better. A few horrible hacks later, I proudly present you "autoresize"
(it was almost named "odooresize", but common sense prevailed)

Note that it also has the benefit of working better with the web client.
For example, when you open a dialog in a model, the autosize fields
should have the correct height automagically (or should I say
'odoomagically')