[FIX] web_gantt: "Invalid day index" when start==end
Library dhtmlxGantt is used to parse and display task.

When a task is parsed, if start_date equals end_date, the end_date is
not kept (there is an boolean "$no_end" variable indicating it), but
this is not taken into account Odoo side and later when getting the
position of the "undefined" end date an error "Invalid day index" is
displayed.

This commit improves this by using start_date in this instance.

closes #191
opw-665441