[FIX] web: change menu section after 'home action' loading
The 'home action' field allows to specify which action to load at log for each
user. This was still working fine in the new webclient, except that the menu
corresponding to the given action wasn't displayed. We thus force a change
menu section once the 'home action' is loaded.

The heuristic to determine when the 'home action' should be executed relies on
the current url: if it is empty, it should mean that we just logged in, so we
execute it. This isn't true anymore since the introduction of the app switcher.
Indeed, the app switcher's url is empty as well, so a refresh on the app
switcher's screen would load the 'home action', which is probably not what we
want. To circumvent this, we set a dummy keyword in the url when the app
switcher is displayed.

opw 657952