[FIX] web: m2m tag width when large text
When a m2m tag badge has a text which is larger than the group is in,
it forces the group to grow (and make it break over the next line
sometimes).

Perfect solution would be to limit the badge width to the group one but
unfortunatly this is not possible (table cells always grow if their
contents are larger and % sizes for their contents do apply after the
initial cell widths are computed). Solution here is to limit the badge
width to a fixed px width.