[FIX] website_enterprise: correctly redirect to app switcher
When the user has a 'home action' defined, clicking on the app switcher button
in the website didn't correctly redirect to the app switcher, as it rather
load the 'home action' (because the url was empty). This rev. simply forces
the app switcher to be displayed by adding '#home' in the url, which is the
exact url of the app switcher.