[FIX] account_contract_dashboard: missing translation
In the dashboard, the title of each graph is not translated.

The dashboard works as follow:
- the JS widget (e.g. account_contract_dashboard_main) calls a
  controller
- the controller retrieves the values to display, and sends them back as
  a dictionary
- the JS widget renders the appropriate template using the data
  retrieved from the controller

This implies that in the JS layer expects these data to be already
translated. Indeed, the JS layer only receives some specific type of
terms all terms (see file translate.py for details, e.g. variable
TRANSLATED_ATTRS or terms located in '*.js' files).

The fix fetches the translation at the level of the controller before
sending the data to the JS layer.

opw-666577