[FIX] sale_contract: change the recurrency when changing the template
When switching template, the recurrency should change even if
the susbcription is already present in the database.

Previously, only newly created record not yet saved had this
behaviour. This makes sense for the lines (you don't want to
clean the lines everytime you change the template, and a smart
uom-conversion function was not implemented) but not for the
recurrency, especially since the recurrency is no longer visible
in the form view.