[IMP] static: cancel table-responsive when printing
Printing with CTRL+P, a table on a page can benefit from a feature
called shrinking for example in google chrome. This allow for example
for a too big table to be shrinked (ofter with a max shrinking factor)
so when the table is printed we see more than we would have with the
initial aspect ratio.

But when "table-responsive" bootstrap class is used, it would not allow
this shrinking since the overflow would not be shown, so there is no
need to shrink to show it.

This commit cancel the .table-responsive class when printing for such
instances.

closes #202

opw-659430
opw-660763