[FIX] web: load async js modules as soon as possible
The module system allow modules to return a deferred.  When that
deferred resolves, the module is considered to have succesfully loaded,
and the result value is its value.  Then, the boot.js code used a
setTimeout (via _.defer) to load all its dependencies.

This is not wrong per se, but it empties the call stack, which give
other code the opportunity to interfere.  In particular, it is an issue
in phantomjs tests, because it causes a window between the
domcontentloaded event (which is a key event in website) and the actual
loading of all dependent modules.  This execution window can be used to
start some tours, but the code of those tours is waiting the next tick
to load, so horrible crashes happens.

This commit calls immediately process_jobs to avoid those issues.