[FIX] project_timesheet_synchro : clean timesheet app xml ids
Some records created by the timesheet app had an xml_id starting with the name of the module.
This means that upon upgrade or uninstall the records would be deleted, which is not the desired behavior.

To prevent this problem from occuring, this commit adds a method in the backend to convert the guilty xml ids.
The method will be called from the app at sync. If this method is not available and problematic ids are present, a warning is shown to the user.
Once the ids have been cleaned in DB, the app will detect it and clean the locally stored records, ensuring appropriate sync between db and external device.