[FIX] web: improve focus/blur one2many dataset
The dataset of an x2m could be set as changed on a focus event, but if
the form is currently not in editing, it could stay blocked in this
state.

This commit add a check so blur and focus on x2m are better coupled.

closes #10637
opw-665401
community: odoo/odoo@b55e29e