[FIX] web: listview editable, non editable elements
When editing a line, leave the buttons and the readonly cells visible.
Improves the style of readonly elements to keep the padding and colors
of the cell below.

Also remove the 3px gap between fields so that readonly field texts
does not move when entering the edition mode. Also leave the class
'o_form_field_number' on the field in readonly so that the listview
can align them to the right.

Notice that chrome still has a bug which makes the fields in edition
go 1px too high for the first line (this is due to the border-bottom
of the thead element ...).