[FIX] voip: update the sipjs lib to fix issues with the new Asterisk, Chrome and Firefox versions
forward port of 1d692b7d10d9ebfc8051ae11ebdf7998ff460474
New versions of Chrome and Firefox introduced a new way of managing the media
streams, which doesn't work with older version of Asterisk.
We keep the old lib to allow customers to still use it they are still using
the old versions of Asterisk, Chrome and Firefox, and they don't upgrade voip.