[FIX] currency_rate_live: change of company currency
When changing the company currency,
e.g. from USD to EUR,
and hitting the "Update now" button on the currency
update feature, while selecting Yahoo as provider,
the company currency rate must be set to 1.000,
as the currencies rate will be fetched against
the company currency.
Besides, this is stated in the method documentation:
```
This method is used to update the currencies by using Yahoo service provider.
Rates are given against the company currency, which will be also updated to a rate of 1
```

opw-668143