[IMP] pos_blackbox_be: consolidate orderlines when allowed
Every orderline has to be registered in the EJ and FDM. After that
you're allowed to consolidate them. So in our case we will consolidate
after creation of a bill.

So what will happen is:
1. send unconsolidated to EJ and FDM
2. print unconsolidated bill
3. consolidate

In theory we could extend this consolidation to happen after implicit
creation of a pro forma order by switching between tables or
orders. Although this would be legal I think it would be too confusing
for the user if when he switches to a different order and then switches
back he notices that his order has changed. This solution is just as
legal, and easier to understand and explain.