[FIX] website_contract: code in uncalled function is not called
Probably a salepocalypse problem, but SOL don't have a 'button_confirm'
so call this in the order confirmation instead.

As long as we're here, the chatter message should be set on the subscription,
not on the analytic account.

Leftover from v8~saas6: use the field subscription_id for
sale subscription linked to so, not project_id.