[FIX] website_sign: requests can be draft
a `signature.request` can be draft,
`draft` is wihtin the possible values for
the `state` selection field.

The function `prepare_requests` is called
using the result of the method `get_dashboard_info`,
which doesn't use a domain excluding
the draft requests:
```
    def get_dashboard_info(self):
            recordset = self.search([])
```
This is therefore very possible to call `prepare_requests`
with draft requests.

opw-668866