[FIX] web: pivot click keep view references
sideport of commit d597463493f76af9d3a4fe7243fa2a12b220b410 from
community by nle.

When clicking on a pivot view cell, the gotten view were default views.

For example when on the opportunity pivot view, after clicking on a cell
we we would get to the crm.lead default view instead of the opportunity
one.

This commit keep references to current action views when clicking on the
cell.