[FIX] sale_contract: deletion cascade for subscription lines
When deleting a subscription, you expect subscription lines to
be deleted as well. If not, you can't delete a product linked
to a subscription line (even if you can't find it anymore since
it's not longer linked to a subscription...).

Even though this is a model change, @KangOl gave me his permission
to commit this in stable since it is retrocompatible (nothing happens
if you don't -u and you enjoy the full effect of this fix for a new
db or when you -u your db).