[FIX] web_gantt: set parameter round_dnd_dates
dhtmlxgantt expects a boolean value for round_dnd_dates
Previous code would apply the string on the gantt config instead of the boolean
value.
Casting the value to boolean is not a good idea too as `"false"` would be casted
to `true`.

Side note: it is not possible to set the flag to false but it's the default
value.