[FIX] web: avoid traceback on register if mail is not installed
The registration process depends on mail through the update notification call.
Instead of an ugly traceback if mail is not installed, show an error notification
instead. Since using a db without mail is not gonna get you very far, we can assume
that most installs will have it install at the moment of registration. For the few
cases where someone wants to register an empty db, this warning is more elegant.