[IMP] web: view_manager: lazy loading of default multi-record view
After a refresh or when opening odoo in a new tab, if view_type="form" is
specified in the url, RPCs to load the default multi-record view will be
performed if/when the user clicks on the breadcrumbs.

Moreover, the view manager is directly initialized with the res_id to load
after a refresh on a form view, to prevent the form view to fetch the defaults
values, to render in 'create' more, and then load the requested record and
switch back to 'view' mode.

Some code refactoring as well:
  - don't propagate state to the view_manager, but directly the view_type and
    res_id
  - remove unused argument no_store from switch_mode()
  - view_registry.get(): remove unused second argument
  - no need to switch_mode in do_load_state() as the correct mode is now loaded
    directly, even if it is a mono-record view