[FIX] web: listview editable elements positioning
Commit 3710e98321743843e9c8b65fad596826d3f23d1c makes the translate icon position computed alongside the
element position computation and changes this to be on the right top of
the cell. The problem is that for some widgets (image, checkbox, ...),
the position had to be the left top of the cell as these do not have a
width which is 100% of the cell.

This also fixes a bug for image fields : on the first edition of a line
which contains such a field, the image width is unknown when computing
the element position (as it is not loaded yet) and so placing it on the
right top made it appear on the next cell (as the width was 0 for the
computation algo). For left top positioning, the width of the element
does not matter.